[ISSUE-63] [PERFORMANCE] Improve DesktopNavMenu. Make it lazyLoaded. Since chakra-Menu component wights like 20kb
