+++ components/Header/DesktopHeader.tsx
@@ -3,12 +3,46 @@ import {
   Center,
   HStack,
   Heading,
+  Link,
+  LinkProps
 } from "@chakra-ui/react";
 import MENU_LINKS from "./navLinks";
-import DesktopNavMenu, { DesktopNavLink } from "./DesktopNavMenu";
+
 import NextLink from "next/link";
+import dynamic from "next/dynamic";
+import { useCallback, useState } from "react";
+import { trackEvent } from "@/lib/gtag";
+import { INavLink } from "@/types/types";
+
+export interface DesktopNavLinkProps extends LinkProps, INavLink{}
+
+export const DesktopNavLink = ({link, label, isMenu, ...rest}: DesktopNavLinkProps) => {
+  const onClickAction = useCallback(() => {
+    trackEvent("desktop_nav",  { link })
+  },[link])
+  return (
+    <NextLink href={link} passHref>
+      <Link
+        {...rest}
+        onClick={onClickAction}
+      >
+        {label}
+      </Link>
+    </NextLink>
+  );
+}
+
+const DesktopNavMenu = dynamic(
+  () => import("./DesktopNavMenu")
+);
 
 const DesktopHeader = () => {
+  const [ menuInstaciated, setMenuInstaciated] = useState(false);
+
+  const openMenu = useCallback(() => {
+    setMenuInstaciated(true);
+  }, [])
+  
   return (
     <Flex justify={"space-between"} width="100%" height={20} as="header">
       <Center w={"300px"} alignContent={"center"} justifyContent={"flex-start"}>
@@ -25,7 +59,7 @@ const DesktopHeader = () => {
         {MENU_LINKS &&
           MENU_LINKS.map((item) => {
             if (item.isMenu) {
-              return <DesktopNavMenu menu={item} key={item.label} />;
+              return menuInstaciated ? <DesktopNavMenu menu={item} key={item.label}/> : <Link onClick={openMenu} key={item.label}>{item.label}</Link>;
             } else {
               return <DesktopNavLink {...item} key={item.label} />;
             }
