+++ components/Header/DesktopNavMenu.tsx
@@ -1,33 +1,16 @@
 // import { ChevronDownIcon } from "@chakra-ui/icons";
-import { Menu, MenuButton, MenuList, MenuItem, Link, LinkProps } from "@chakra-ui/react";
+import { Menu, MenuButton, MenuList, MenuItem, Link, LinkProps, useDisclosure } from "@chakra-ui/react";
 import { INavMenu } from "@/types/types";
 import { INavLink } from "@/types/types";
+import { useState } from "react";
+import { DesktopNavLink } from "./DesktopHeader";
 
-import NextLink from "next/link"
-import { trackEvent } from "@/lib/gtag";
-import { useCallback } from "react";
 
-export interface DesktopNavLinkProps extends LinkProps, INavLink{}
+const DesktopNavMenu: React.FC<{ menu: INavMenu }> = ({ menu }) => {
+  const [openState, setOpenState] = useState(true);
 
-export const DesktopNavLink = ({link, label, isMenu, ...rest}: DesktopNavLinkProps) => {
-  const onClickAction = useCallback(() => {
-    trackEvent("desktop_nav",  { link })
-  },[link])
   return (
-    <NextLink href={link} passHref>
-      <Link
-        {...rest}
-        onClick={onClickAction}
-      >
-        {label}
-      </Link>
-    </NextLink>
-  );
-}
-
-const DesktopNavMenu: React.FC<{ menu: INavMenu }> = ({ menu }) => (
-  //TODO Menu component is INSANLY BIG. weigths like 50kb. Should replace it with some ad-hoc component o just lazy load it when the user hover on the link
-  <Menu colorScheme={"pink"}>
+  <Menu colorScheme={"pink"} isLazy={openState} isOpen={openState} onClose={()=> setOpenState(false)} onOpen={() => setOpenState(true)}>
     <MenuButton
       as={Link}
       _focus={{ boxShadow: "none" }}
@@ -43,6 +26,6 @@ const DesktopNavMenu: React.FC<{ menu: INavMenu }> = ({ menu }) => (
       ))}
     </MenuList>
   </Menu>
-);
+)};
 
 export default DesktopNavMenu;
